requirements.txt
streamlit
validators
langchain
langchain_groq
langchain_community
python-dotenv
yt-dlp
youtube-transcript-api
pytube
unstructured
